requirements.txt
torch>=2.0.0
transformers>=4.36.0
Pillow>=9.0.0
matplotlib>=3.5.0
requests>=2.25.0
numpy>=1.21.0
opencv-python
accelerate>=0.25.0
safetensors>=0.4.0